# Speed up `keys` by using foldrWithIndex