--- a/bench/Bench/Data/Map.purs
+++ b/bench/Bench/Data/Map.purs
@@ -46,7 +46,7 @@ benchMap = do
 
   log "keys"
   log "---------------"
-  benchValues
+  benchKeys
 
   where