**Description of the change**

Addresses the comment raised by @xgrommx here: https://github.com/purescript/purescript-ordered-collections/pull/61#issuecomment-1106612371=
```
Map
===
keys
---------------
Map2a0bff.keys: big map (1000000)
mean   = 2.64 s
stddev = 199.37 ms
min    = 2.39 s
max    = 2.96 s
M.keys: big map (1000000)
mean   = 218.74 ms
stddev = 34.92 ms
min    = 194.87 ms
max    = 301.29 ms
```

---

**Checklist:**

- [ ] Added the change to the changelog's "Unreleased" section with a reference to this PR (e.g. "- Made a change (#0000)")
- [ ] Linked any existing issues or proposals that this pull request should close
- [ ] Updated or added relevant documentation
- [ ] Added a test for the contribution (if applicable)